database-service/requirements.txt
argon2-cffi==19.1.0
flask==1.1.1
pymysql==0.9.3
requests==2.22.0
SQLAlchemy==1.3.7
